+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/AsyncVoiceMessages/VoiceRecordingContainerView.swift
@@ -91,7 +91,11 @@ public struct VoiceRecordingContainerView<Factory: ViewFactory>: View {
         .background(Color(colors.background))
         .cornerRadius(16)
         .padding(.all, 4)
-        .messageBubble(for: message, isFirst: isFirst)
+        .modifier(
+            factory.makeMessageViewModifier(
+                for: MessageModifierInfo(message: message, isFirst: isFirst)
+            )
+        )
     }
     
     private func index(for attachment: ChatMessageVoiceRecordingAttachment) -> Int {